+++ packages/falcon-client/src/service/ApolloClientWeb.js
@@ -2,7 +2,7 @@ import { CachePersistor } from 'apollo-cache-persist';
 import { InMemoryCache } from 'apollo-cache-inmemory';
 import { ApolloClient } from './';
 
-export const apolloClientWeb = ({ initialState, clientApolloSchema, apolloClientConfig }) => {
+export const apolloClientWeb = ({ initialState, graphqlProxy, clientApolloSchema, apolloClientConfig }) => {
   const cache = new InMemoryCache({ addTypename: false });
   const cachePersistor = new CachePersistor({
     cache,
@@ -21,6 +21,7 @@ export const apolloClientWeb = ({ initialState, clientApolloSchema, apolloClient
     const apolloClient = new ApolloClient({
       isBrowser: true,
       clientState: clientApolloSchema,
+      graphqlProxy,
       initialState,
       apolloClientConfig,
       cache